add custom type example
